@@ -1,6 +1,6 @@
 {
   "name": "nx-console",
-  "version": "9.0.0",
+  "version": "17.9.0",
   "publisher": "nrwl",
   "author": {
     "name": "Narwhal Technologies Inc",
@@ -26,47 +26,59 @@
     "update": "nx migrate latest"
   },
   "dependencies": {
-    "@ngrx/component-store": "11.0.0",
-    "jsonc-parser": "^2.1.1",
-    "strip-json-comments": "^3.1.1"
+    "jsonc-parser": "^3.0.0"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "0.1102.1",
-    "@angular/cli": "11.2.1",
-    "@angular/common": "11.2.1",
-    "@angular/compiler": "11.2.1",
-    "@angular/compiler-cli": "11.2.1",
-    "@angular/core": "11.2.1",
-    "@angular/forms": "11.2.1",
-    "@angular/platform-browser": "11.2.1",
-    "@angular/platform-browser-dynamic": "11.2.1",
+    "@angular-devkit/build-angular": "12.1.4",
+    "@angular-eslint/eslint-plugin": "12.3.1",
+    "@angular-eslint/eslint-plugin-template": "12.3.1",
+    "@angular-eslint/template-parser": "12.3.1",
+    "@angular/cdk": "11.2.1",
+    "@angular/cli": "12.1.4",
+    "@angular/common": "12.2.2",
+    "@angular/compiler": "12.2.2",
+    "@angular/compiler-cli": "12.2.2",
+    "@angular/core": "12.2.2",
+    "@angular/forms": "12.2.2",
+    "@angular/platform-browser": "12.2.2",
+    "@angular/platform-browser-dynamic": "12.2.2",
     "@babel/core": "7.5.4",
-    "@nrwl/angular": "11.4.0",
-    "@nrwl/cli": "11.4.0",
-    "@nrwl/cypress": "11.4.0",
-    "@nrwl/jest": "11.4.0",
-    "@nrwl/node": "11.4.0",
-    "@nrwl/nx-cloud": "11.1.4",
-    "@nrwl/storybook": "11.4.0",
-    "@nrwl/tao": "11.4.0",
-    "@nrwl/workspace": "11.4.0",
-    "@storybook/addon-knobs": "^6.1.11",
-    "@storybook/angular": "^6.1.11",
+    "@nrwl/angular": "12.8.0",
+    "@nrwl/cli": "12.8.0",
+    "@nrwl/cypress": "12.8.0",
+    "@nrwl/eslint-plugin-nx": "12.8.0",
+    "@nrwl/jest": "12.8.0",
+    "@nrwl/linter": "12.8.0",
+    "@nrwl/node": "12.8.0",
+    "@nrwl/nx-cloud": "12.3.10",
+    "@nrwl/storybook": "12.8.0",
+    "@nrwl/tao": "12.8.0",
+    "@nrwl/workspace": "12.8.0",
+    "@storybook/addon-essentials": "~6.3.0",
+    "@storybook/addon-knobs": "~6.3.0",
+    "@storybook/angular": "~6.3.0",
+    "@storybook/builder-webpack5": "~6.3.0",
+    "@storybook/manager-webpack5": "~6.3.0",
     "@types/jest": "26.0.8",
-    "@types/node": "12.12.38",
+    "@types/node": "14.14.33",
     "@types/shelljs": "^0.8.2",
     "@types/universal-analytics": "0.4.2",
     "@types/uuid": "^3.4.5",
-    "babel-loader": "8.0.6",
-    "codelyzer": "6.0.1",
+    "@types/vscode": "1.50.0",
+    "@typescript-eslint/eslint-plugin": "4.28.5",
+    "@typescript-eslint/parser": "4.28.5",
+    "copy-webpack-plugin": "^9.0.1",
     "cypress": "6.5.0",
-    "fast-glob": "3.0.4",
+    "eslint": "7.10.0",
+    "eslint-config-prettier": "8.1.0",
     "husky": "^3.0.9",
-    "jest": "26.2.2",
-    "jest-preset-angular": "8.3.2",
+    "jest": "27.0.3",
+    "jest-preset-angular": "9.0.3",
+    "ngx-build-plus": "11.0.0",
     "nps": "5.9.3",
     "nps-utils": "1.7.0",
-    "prettier": "2.2.1",
+    "ovsx": "^0.2.0",
+    "prettier": "2.3.2",
     "pretty-quick": "^1.11.1",
     "rxjs": "6.5.5",
     "shelljs": "0.8.3",
@@ -75,16 +87,16 @@
     "standard-version": "^9.1.1",
     "stylelint": "^10.1.0",
     "stylelint-config-standard": "^18.3.0",
-    "ts-jest": "26.4.0",
-    "tsickle": "0.35.0",
+    "ts-jest": "27.0.3",
     "tslib": "1.14.1",
-    "tslint": "6.1.3",
-    "tslint-no-unused-expression-chai": "^0.1.4",
-    "typescript": "4.0.7",
+    "tslint": "^6.1.3",
+    "typescript": "4.3.5",
     "universal-analytics": "0.4.20",
     "vsce": "^1.69.0",
-    "vscode": "^1.1.35",
-    "zlib": "^1.0.5",
-    "zone.js": "0.10.3"
+    "vscode-codicons": "^0.0.16",
+    "webpack": "^5.47.0",
+    "webpack-merge": "^5.8.0",
+    "webpack-node-externals": "^3.0.0",
+    "zone.js": "~0.11.4"
   }
 }
