@@ -1,7 +1,6 @@
-/* tslint:disable */
 
 import * as ua from 'universal-analytics';
-const uuidv4 = require('uuid/v4');
+import uuidv4 = require('uuid/v4');
 
 export class Telemetry {
   private readonly visitor = new ua.Visitor('UA-88380372-8', this.getUuiId(), {
