@@ -1,5 +1,3 @@
-import { TASK_EXECUTION_SCHEMA } from '@nx-console/vscode-ui/feature-task-execution-form';
-
 // This file can be replaced during build by using the `fileReplacements` array.
 // `ng build --prod` replaces `environment.ts` with `environment.prod.ts`.
 // The list of file replacements can be found in `angular.json`.
@@ -270,8 +268,7 @@ export const NG_BUILD_ARCHITECT = {
     {
       name: 'buildOptimizer',
       type: 'boolean',
-      description:
-        "Enables '@angular-devkit/build-optimizer' optimizations when using the 'aot' option.",
+      description: `Enables '@angular-devkit/build-optimizer' optimizations when using the 'aot' option.`,
       defaultValue: 'false',
       required: false,
       positional: false,
@@ -325,8 +322,7 @@ export const NG_BUILD_ARCHITECT = {
     {
       name: 'statsJson',
       type: 'boolean',
-      description:
-        "Generates a 'stats.json' file which can be analyzed using tools such as 'webpack-bundle-analyzer'.",
+      description: `Generates a 'stats.json' file which can be analyzed using tools such as 'webpack-bundle-analyzer'.`,
       defaultValue: 'false',
       required: false,
       positional: false,
@@ -700,7 +696,7 @@ export const MOCK_COMPONENT_ARCHITECT = {
       enum: null,
       type: 'boolean',
       description:
-        'When true (the default), generates a "spec.ts" test file for the new component.',
+        'When true (the default), generates a `spec.ts` test file for the new component.',
       defaultValue: true,
       required: false,
       positional: false,
@@ -712,7 +708,7 @@ export const MOCK_COMPONENT_ARCHITECT = {
       enum: null,
       type: 'boolean',
       description:
-        'When true, does not create "spec.ts" test files for the new component.',
+        'When true, does not create `spec.ts` test files for the new component.',
       defaultValue: false,
       required: false,
       positional: false,
@@ -783,10 +779,5 @@ export const MOCK_COMPONENT_ARCHITECT = {
 
 export const environment = {
   production: false,
-  providers: [
-    {
-      provide: TASK_EXECUTION_SCHEMA,
-      useValue: MOCK_COMPONENT_ARCHITECT,
-    },
-  ],
+  providers: [],
 };
