@@ -3,7 +3,7 @@ import { platformBrowserDynamic } from '@angular/platform-browser-dynamic';
 
 import { AppModule } from './app/app.module';
 import { environment } from './environments/environment';
-import 'zone.js/dist/zone';
+import 'zone.js';
 
 if (environment.production) {
   enableProdMode();
