@@ -1,8 +1,5 @@
+const { getJestProjects } = require('@nrwl/jest');
+
 module.exports = {
-  projects: [
-    '<rootDir>/apps/vscode-ui',
-    '<rootDir>/libs/server',
-    '<rootDir>/libs/vscode-ui/components',
-    '<rootDir>/libs/vscode-ui/feature-task-execution-form',
-  ],
+  projects: getJestProjects(),
 };
