@@ -1,7 +1,7 @@
 module.exports = {
   preset: '../../jest.preset.js',
   coverageDirectory: '../../coverage/libs/server',
-  setupFilesAfterEnv: ['<rootDir>/src/test-setup.ts'],
-  globals: { 'ts-jest': { tsConfig: '<rootDir>/tsconfig.spec.json' } },
+  globals: { 'ts-jest': { tsconfig: '<rootDir>/tsconfig.spec.json' } },
   displayName: 'server',
+  testEnvironment: 'node',
 };
