@@ -1,9 +1,11 @@
-import { Component, ViewChild } from '@angular/core';
+import { Component, DebugElement, ViewChild } from '@angular/core';
 import { waitForAsync, ComponentFixture, TestBed } from '@angular/core/testing';
+import { By } from '@angular/platform-browser';
 import { FormBuilder, FormsModule, ReactiveFormsModule } from '@angular/forms';
 import { OptionType } from '@angular/cli/models/interface';
 import { Option } from '@nx-console/schema';
 import { SelectComponent } from './select.component';
+import { FieldItemsPipe } from '../field-items/field-items.pipe';
 
 const initialValue = 'test';
 const mockOption: Option = {
@@ -25,8 +27,15 @@ const mockOption: Option = {
   `,
 })
 class ParentFormComponent {
-  field = mockOption;
-  formGroup = this.fb.group({ [this.field.name]: initialValue });
+  formGroup = this.fb.group({ [mockOption.name]: initialValue });
+  _field = mockOption;
+  get field(): Option {
+    return this._field;
+  }
+  set field(field: Option) {
+    this._field = field;
+    this.formGroup = this.fb.group({ [field.name]: '' });
+  }
   @ViewChild('select', { static: true })
   selectComponent: SelectComponent;
 
@@ -35,11 +44,12 @@ class ParentFormComponent {
 describe('SelectComponent', () => {
   let fixture: ComponentFixture<ParentFormComponent>;
   let parent: ParentFormComponent;
+  let options: DebugElement[];
 
   beforeEach(
     waitForAsync(() => {
       TestBed.configureTestingModule({
-        declarations: [ParentFormComponent, SelectComponent],
+        declarations: [ParentFormComponent, SelectComponent, FieldItemsPipe],
         imports: [FormsModule, ReactiveFormsModule],
       }).compileComponents();
     })
@@ -48,6 +58,7 @@ describe('SelectComponent', () => {
     fixture = TestBed.createComponent(ParentFormComponent);
     parent = fixture.componentInstance;
     fixture.detectChanges();
+    options = fixture.debugElement.queryAll(By.css('option'));
   });
 
   it('should create', () => {
@@ -55,16 +66,39 @@ describe('SelectComponent', () => {
     expect(parent.selectComponent).toBeDefined();
   });
 
-  describe('getOptionTooltip', () => {
+  describe('option tooltips', () => {
     it('should get tooltip for option value', () => {
-      parent.field = mockOption;
-      expect(parent.selectComponent.getOptionTooltip('test')).toBe('testLabel');
+      const testOption = options.find(
+        (opt) => opt.nativeElement.textContent.trim() === 'test'
+      );
+      expect(testOption).toBeDefined();
+      expect(testOption?.nativeElement.getAttribute('title')).toEqual(
+        mockOption.itemTooltips && mockOption.itemTooltips['test']
+      );
     });
-    it('should return null when there is no tooltip for option value', () => {
-      parent.field = mockOption;
-      expect(
-        parent.selectComponent.getOptionTooltip('noTooltipValue')
-      ).toBeNull();
+    it('should ignore when there is no tooltip for option value', () => {
+      const noTooltipOption = options.find(
+        (opt) => opt.nativeElement.textContent.trim() === 'other'
+      );
+      expect(noTooltipOption).toBeDefined();
+      expect(noTooltipOption?.nativeElement.getAttribute('title')).toBeNull();
     });
+  });
+
+  it('should show long form options', () => {
+    const longForm = {
+      name: 'option-items-with-enum',
+      type: OptionType.String,
+      aliases: [],
+      description: 'a long form select option',
+      default: 'scss',
+      items: {
+        type: OptionType.String,
+        enum: ['css', 'scss', 'styl', 'less'],
+      },
+    };
+    parent.field = longForm;
+    fixture.detectChanges();
+    expect(fixture.debugElement.queryAll(By.css('#option-items-with-enum option')).length).toEqual(4);
   });
 });
