@@ -3,7 +3,6 @@ import {
   Component,
   EventEmitter,
   Input,
-  OnInit,
   Output,
 } from '@angular/core';
 import { ControlContainer, FormGroup } from '@angular/forms';
@@ -15,30 +14,18 @@ import { Option } from '@nx-console/schema';
   styleUrls: ['./select.component.scss'],
   changeDetection: ChangeDetectionStrategy.OnPush,
 })
-export class SelectComponent implements OnInit {
+export class SelectComponent {
   @Input() field: Option;
   @Input() value: string;
   @Output() readonly valueChange = new EventEmitter<string>();
-  parentFormGroup: FormGroup;
+
+  get parentFormGroup(): FormGroup {
+    return this.controlContainer.control as FormGroup;
+  }
 
   constructor(private readonly controlContainer: ControlContainer) {}
 
-  ngOnInit() {
-    this.parentFormGroup = this.controlContainer.control as FormGroup;
-  }
-
   updateValue(updated: string): void {
     this.valueChange.emit(updated);
-  }
-
-  getOptionTooltip(optionValue: string): string | null {
-    if (
-      this.field &&
-      this.field.itemTooltips &&
-      this.field.itemTooltips[optionValue]
-    ) {
-      return this.field.itemTooltips[optionValue];
-    }
-    return null;
   }
 }
