@@ -8,7 +8,7 @@ import { CheckboxComponent } from './checkbox/checkbox.component';
 import { FieldComponent } from './field/field.component';
 import { FieldTreeComponent } from './field-tree/field-tree.component';
 import { MultipleSelectComponent } from './multiple-select/multiple-select.component';
-import { FormatTaskPipe } from './format-task/format-task.pipe';
+import { FieldItemsPipe } from './field-items/field-items.pipe';
 
 @NgModule({
   imports: [CommonModule, ReactiveFormsModule],
@@ -20,7 +20,7 @@ import { FormatTaskPipe } from './format-task/format-task.pipe';
     FieldComponent,
     FieldTreeComponent,
     MultipleSelectComponent,
-    FormatTaskPipe,
+    FieldItemsPipe,
   ],
   exports: [
     AutocompleteComponent,
@@ -30,7 +30,6 @@ import { FormatTaskPipe } from './format-task/format-task.pipe';
     FieldComponent,
     FieldTreeComponent,
     MultipleSelectComponent,
-    FormatTaskPipe,
   ],
 })
 export class VscodeUiComponentsModule {}
