@@ -1,1 +1,1 @@
-import 'jest-preset-angular';
+import 'jest-preset-angular/setup-jest';
