@@ -5,6 +5,8 @@ import { waitForAsync, ComponentFixture, TestBed } from '@angular/core/testing';
 import { TaskExecutionSchema } from '@nx-console/schema';
 import { TASK_EXECUTION_SCHEMA } from './task-execution-form.schema';
 import { VscodeUiComponentsModule } from '@nx-console/vscode-ui/components';
+import { ArgumentListModule } from '@nx-console/vscode-ui/argument-list';
+import { FormatTaskPipe } from './format-task/format-task.pipe';
 import { TaskExecutionFormComponent } from './task-execution-form.component';
 
 const initialSchema: TaskExecutionSchema = {
@@ -31,6 +33,16 @@ const initialSchema: TaskExecutionSchema = {
       description: 'a multiselect option',
       items: ['one', 'two', 'three', 'four'],
     },
+    {
+      name: 'a-long-form-multiselect-option',
+      type: OptionType.Array,
+      aliases: [],
+      description: 'a long form multiselect option',
+      items: {
+        type: OptionType.String,
+        enum: ['five', 'six', 'seven', 'eight'],
+      },
+    },
   ],
 };
 
@@ -42,8 +54,12 @@ describe('TaskExecutionFormComponent', () => {
   beforeEach(
     waitForAsync(() => {
       TestBed.configureTestingModule({
-        declarations: [TaskExecutionFormComponent],
-        imports: [ReactiveFormsModule, VscodeUiComponentsModule],
+        declarations: [TaskExecutionFormComponent, FormatTaskPipe],
+        imports: [
+          ReactiveFormsModule,
+          VscodeUiComponentsModule,
+          ArgumentListModule,
+        ],
         providers: [
           { provide: TASK_EXECUTION_SCHEMA, useValue: initialSchema },
         ],
@@ -62,21 +78,36 @@ describe('TaskExecutionFormComponent', () => {
     expect(component).toBeTruthy();
   });
 
-  it('should set validator for valid options', () => {
-    const formControl = formGroup.get('option-items-with-enum') as FormControl;
-    expect(formControl?.validator).toBeDefined();
-    formControl.setValue('not a valid option');
-    expect(formControl.invalid).toBeTruthy();
-    formControl.setValue('scss');
-    expect(formControl.valid).toBeTruthy();
-  });
+  describe('validators', () => {
+    it('should set validator for valid options', () => {
+      const formControl = formGroup.get(
+        'option-items-with-enum'
+      ) as FormControl;
+      expect(formControl?.validator).toBeDefined();
+      formControl.setValue('not a valid option');
+      expect(formControl.invalid).toBeTruthy();
+      formControl.setValue('scss');
+      expect(formControl.valid).toBeTruthy();
+    });
 
-  it('should set validator for Array type multiselect options', () => {
-    const formControl = formGroup.get('a-multiselect-option') as FormControl;
-    expect(formControl?.validator).toBeDefined();
-    formControl.setValue(['not valid']);
-    expect(formControl.invalid).toBeTruthy();
-    formControl.setValue(['one', 'three']);
-    expect(formControl.valid).toBeTruthy();
+    it('should set validator for Array type multiselect options', () => {
+      const formControl = formGroup.get('a-multiselect-option') as FormControl;
+      expect(formControl?.validator).toBeDefined();
+      formControl.setValue(['not valid']);
+      expect(formControl.invalid).toBeTruthy();
+      formControl.setValue(['one', 'three']);
+      expect(formControl.valid).toBeTruthy();
+    });
+
+    it('should set validator for long form options', () => {
+      const formControl = formGroup.get(
+        'a-long-form-multiselect-option'
+      ) as FormControl;
+      expect(formControl?.validator).toBeDefined();
+      formControl.setValue(['not valid']);
+      expect(formControl.invalid).toBeTruthy();
+      formControl.setValue(['five', 'six']);
+      expect(formControl.valid).toBeTruthy();
+    });
   });
 });
